requirements.txt
-f /usr/share/pip-wheels
aioredis==1.3.1
amqp==2.6.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
aws==0.2.5
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
billiard==3.6.4.0
boto==2.49.0
boto3==1.13.24
botocore==1.16.24
Brotli==1.0.9
buildozer==1.2.0
celery==4.4.2
certifi==2020.4.5.1
cfe
cffi==1.14.0
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography
cssselect2==0.4.1
daphne==3.0.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.42.0
django-bootstrap-markdown-editor==1.3.1
django-bootstrap-v5==1.0.6
django-bootstrap4==2.1.1
django-bootstrap5==2.1.2
django-braces==1.14.0
django-cors-headers==3.6.0
django-crispy-forms==1.9.1
django-embed-video @ git+https://github.com/jazzband/django-embed-video@692a9c2e3c6c6817eb5b596339abe553b55113e7
django-extensions==3.1.3
django-hashid-field==3.1.3
django-localflavor==3.1
django-markdown-deux==1.0.5
django-oauth-toolkit==1.3.2
django-oauth2-provider==0.2.6.1
django-pagedown==2.1.2
django-phonenumber-field==5.0.0
django-pylibmc==0.6.1
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-smtp-ssl==1.0
django-storages==1.9.1
django-summernote==0.8.11.6
django-taggit==1.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
docutils==0.15.2
fabric==2.5.0
filelock==3.0.12
fonttools==4.27.1
gevent==21.1.0
greenlet==1.0.0
gunicorn==20.0.4
hashids==1.3.1
hiredis==1.1.0
html5lib==1.1
hyperlink==20.0.1
idna==2.9
importlib-metadata==3.10.1
incremental==17.5.0
invoke==1.4.1
jmespath==0.10.0
Kivy==2.0.0
Kivy-Garden==0.1.4
kombu==4.6.11
Markdown==3.3.4
markdown-it-py==1.1.0
markdown2==2.3.9
msgpack==1.0.2
mysql==0.0.2
mysqlclient==1.4.6
oauthlib==3.1.0
olefile==0.46
paramiko==2.7.1
pexpect==4.8.0
phonenumberslite==8.12.9
Pillow==7.1.2
prettytable==0.7.2
prompt-toolkit==3.0.20
psycopg2==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydyf==0.1.1
Pygments==2.7.4
PyHamcrest==2.0.2
PyJWT==1.7.1
pylibmc==1.6.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyphen==0.11.0
python-dateutil==2.8.1
python-decouple==3.3
python-memcached==1.59
python-social-auth==0.3.6
python-stdnum==1.16
python3-openid==3.2.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
service-identity==18.1.0
sh==1.14.1
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
smmap2==3.0.1
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
sshtunnel==0.1.5
storage==0.0.4.3
stripe==2.48.0
tinycss2==1.1.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.4.0
wcwidth==0.2.5
weasyprint==53.3
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.57.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.2.0
zopfli==0.1.8
